changes following comments from pull request 1
