+++ vspace/src/main/java/edu/asu/diging/vspace/config/PersistenceJPAConfig.java
@@ -88,8 +88,6 @@ public class PersistenceJPAConfig{
        properties.setProperty("hibernate.hbm2ddl.auto", "update");
        properties.setProperty(
          "hibernate.dialect", "org.hibernate.dialect.MySQL5Dialect");
-       properties.setProperty(
-  	         "hibernate.show_sql", "true");
         
        return properties;
    }
