+++ vspace/src/main/java/edu/asu/diging/vspace/core/model/display/DisplayType.java
@@ -2,5 +2,5 @@ package edu.asu.diging.vspace.core.model.display;
 
 public enum DisplayType {
 
-    ARROW, ALERT, IMAGE;
+    ARROW, ALERT, IMAGE, MODULE;
 }
