+++ vspace/src/main/webapp/WEB-INF/views/exhibition/space.jsp
@@ -50,13 +50,8 @@ function drawLinks() {
         var link = $('<a></a>');
         link.attr('href', '<c:url value="/exhibit/module/${link.link.module.id}" />');
         
-        if ("${link.type}" == 'ALERT') {
-            var linkDisplay = $('<div class="alert alert-primary" role="alert">');
-        } else if ("${link.type}" == 'IMAGE' && "${link.image}" != '') {
-            var linkDisplay = $('<img id="${link.image.id}" src="<c:url value="/api/image/${link.image.id}" />" />');
-        } else {
-            var linkDisplay = $('<span data-feather="navigation-2" class="flex"></span>');
-        }
+        var linkDisplay = $('<span class="fas fa-book-open"></span>');
+     
         linkDisplay.css('position', 'absolute');
         linkDisplay.css('left', ${link.positionX} + posX);
         linkDisplay.css('top', ${link.positionY} + posY);
